+++ src/main/java/com/example/lostnfound/repository/PostRepo.java
@@ -20,6 +20,6 @@ public interface PostRepo extends JpaRepository<Post, Integer>{
     List<Post> findByUserId(Long userId);
 
     @Query(value = "SELECT * FROM post ORDER BY embedding <=> CAST(:queryEmbedding AS vector) LIMIT :topK", nativeQuery = true)
-    List<Post> findTopKSimilarPosts(@Param("queryEmbedding") float[] queryEmbedding, @Param("topK") int topK);
+    List<Post> findTopKSimilarPosts(@Param("queryEmbedding") float[] queryEmbedding, @Param("topK") Long topK);
 
 }
